Style: PeriodContext에 있는 hook을 파일로 분리 - import 순서 변경
