+++ fe/src/contexts/periodContext.tsx
@@ -22,8 +22,8 @@ type PeriodDispatch = Dispatch<Action>;
 
 const thisDate = new Date();
 
-const PeriodStateContext = createContext<Period | null>(null);
-const PeriodDispatchContext = createContext<PeriodDispatch | null>(null);
+export const PeriodStateContext = createContext<Period | null>(null);
+export const PeriodDispatchContext = createContext<PeriodDispatch | null>(null);
 
 const initialState: Period = {
   date: thisDate,
@@ -77,29 +77,4 @@ function reducer(state: Period, action: Action): Period {
     default:
       throw new Error();
   }
-}
-
-export function usePeriodState() {
-  const state = useContext(PeriodStateContext);
-  if (!state) throw new Error();
-
-  return {
-    checkIn: state.checkInOut.checkIn,
-    checkOut: state.checkInOut.checkOut,
-    date: state.date,
-    periodText: state.periodText,
-  };
-}
-
-export function usePeriodDispatch() {
-  const dispatch = useContext(PeriodDispatchContext);
-  if (!dispatch) throw new Error();
-
-  const setCheckIn = (checkIn: Date) => dispatch({ type: 'SET_CHECK_IN', checkIn });
-  const setCheckOut = (checkOut: Date) => dispatch({ type: 'SET_CHECK_OUT', checkOut });
-  const setDate = (date: Date) => dispatch({ type: 'SET_DATE', date });
-  const resetDate = () => dispatch({ type: 'RESET_DATE', payload: null });
-  const setPeriodText = (text: string) => dispatch({ type: 'SET_TEXT', text });
-
-  return { setCheckIn, setCheckOut, setDate, resetDate, setPeriodText };
 }
